fix: Adjust the file structure
