+++ app/(account)/friend-list.tsx
@@ -1,5 +1,16 @@
 import { useColorTheme } from "@/stores/useColorTheme";
-import { DefaultTheme, PaperProvider, Text, ActivityIndicator, Card, Button, Icon, MD3Colors, FAB, Appbar } from "react-native-paper";
+import {
+  DefaultTheme,
+  PaperProvider,
+  Text,
+  ActivityIndicator,
+  Card,
+  Button,
+  Icon,
+  MD3Colors,
+  FAB,
+  Appbar,
+} from "react-native-paper";
 import DarkThemeColors from "@/constants/DarkThemeColors.json";
 import LightThemeColors from "@/constants/LightThemeColors.json";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
@@ -22,17 +33,17 @@ async function fetchFriendList(): Promise<FriendListItem[]> {
   // return new Promise((res, rej) => setTimeout(() => rej(new Error("simulated error")), 5000));
   return [
     {
-      id: 'abcdef',
-      name: 'Michael Coblenz',
+      id: "abcdef",
+      name: "Michael Coblenz",
       requested: true,
     },
     {
-      id: 'fhhqhr',
-      name: 'Antariksha Ray',
+      id: "fhhqhr",
+      name: "Antariksha Ray",
       requested: false,
     },
     {
-      id: 'qhrbaf',
+      id: "qhrbaf",
       name: "Loris D'Antoni",
       requested: false,
     },
@@ -42,7 +53,7 @@ async function fetchFriendList(): Promise<FriendListItem[]> {
 export default function FriendList() {
   const [loaded] = useFonts({
     Poppins: require("@/assets/fonts/Poppins/Poppins-Regular.ttf"), // eslint-disable-line
-    PoppinsBold: require("@/assets/fonts/Poppins/Poppins-Bold.ttf"),// eslint-disable-line
+    PoppinsBold: require("@/assets/fonts/Poppins/Poppins-Bold.ttf"), // eslint-disable-line
   });
 
   React.useEffect(() => {
@@ -66,7 +77,7 @@ export default function FriendList() {
       backgroundColor: theme.colors.primaryContainer,
       padding: 20,
       fontFamily: "Poppins",
-      display: 'flex',
+      display: "flex",
       flexDirection: "column",
       position: "relative",
       flexGrow: 1,
@@ -90,7 +101,7 @@ export default function FriendList() {
       gap: 8,
     },
     fab: {
-      position: 'absolute',
+      position: "absolute",
       margin: 32,
       right: 0,
       bottom: insets.bottom + 32,
@@ -120,54 +131,67 @@ export default function FriendList() {
     return null;
   }
 
-  return <PaperProvider theme={theme}>
-    <View style={stylesheet.container}>
-      <Stack.Screen name="Friends" options={{ title: "Friends" }} />
-      <Appbar.Header
-        style={{
-          height: 50,
-          backgroundColor: theme.colors.primaryContainer,
-        }}
-      >
-        <Appbar.BackAction onPress={() => router.back()} />
-        <Appbar.Content
-          title="Friends"
-          titleStyle={{
-            fontSize: 18,
+  return (
+    <PaperProvider theme={theme}>
+      <View style={stylesheet.container}>
+        <Stack.Screen name="Friends" options={{ title: "Friends" }} />
+        <Appbar.Header
+          style={{
+            height: 50,
+            backgroundColor: theme.colors.primaryContainer,
           }}
-        />
-      </Appbar.Header>
-      <View style={stylesheet.page}>
-        {friends !== null &&
-          <FlatList
-            data={friends}
-            renderItem={({ item }) => <Card style={stylesheet.friendCard}>
-              <Card.Title title={item.name} />
-              <Card.Actions>
-                <Button mode="text" textColor={MD3Colors.error40}>Remove</Button>
-                <Button buttonColor={theme.colors.primary}>View Profile</Button>
-              </Card.Actions>
-            </Card>}
-            style={{marginBottom: tabBarHeight}}
-            keyExtractor={(item) => item.id}
-          />}
-        {fetchError === null && friends === null &&
-          <ActivityIndicator animating={true} size="large" color={theme.colors.onPrimary} />}
-        {fetchError !== null &&
-          <View style={stylesheet.errorMessage}>
-            <Icon source="alert" color={theme.colors.onError} size={48} />
-            <Text>
-              Loading failed: {fetchError.message}. Try again?
-            </Text>
-          </View>}
-        <FAB
-          style={stylesheet.fab}
-          icon="plus"
-          label="Add friend"
-          color={theme.colors.onPrimary}
-          onPress={() => router.push("/(tabs)/account/add-friend")}
-        />
+        >
+          <Appbar.BackAction onPress={() => router.back()} />
+          <Appbar.Content
+            title="Friends"
+            titleStyle={{
+              fontSize: 18,
+            }}
+          />
+        </Appbar.Header>
+        <View style={stylesheet.page}>
+          {friends !== null && (
+            <FlatList
+              data={friends}
+              renderItem={({ item }) => (
+                <Card style={stylesheet.friendCard}>
+                  <Card.Title title={item.name} />
+                  <Card.Actions>
+                    <Button mode="text" textColor={MD3Colors.error40}>
+                      Remove
+                    </Button>
+                    <Button buttonColor={theme.colors.primary}>
+                      View Profile
+                    </Button>
+                  </Card.Actions>
+                </Card>
+              )}
+              style={{ marginBottom: tabBarHeight }}
+              keyExtractor={(item) => item.id}
+            />
+          )}
+          {fetchError === null && friends === null && (
+            <ActivityIndicator
+              animating={true}
+              size="large"
+              color={theme.colors.onPrimary}
+            />
+          )}
+          {fetchError !== null && (
+            <View style={stylesheet.errorMessage}>
+              <Icon source="alert" color={theme.colors.onError} size={48} />
+              <Text>Loading failed: {fetchError.message}. Try again?</Text>
+            </View>
+          )}
+          <FAB
+            style={stylesheet.fab}
+            icon="plus"
+            label="Add friend"
+            color={theme.colors.onPrimary}
+            onPress={() => router.push("/(account)/add-friend")}
+          />
+        </View>
       </View>
-    </View>
-  </PaperProvider>;
+    </PaperProvider>
+  );
 }
