--- app/(tabs)/account/_layout.tsx
@@ -1,8 +1,0 @@
-import { Stack } from "expo-router/stack";
-import React from "react";
-
-export default function Layout() {
-  return <Stack screenOptions={{
-    headerShown: false,
-  }} />
-}
