+++ package.json
@@ -15,12 +15,12 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^14.0.4",
-    "@react-native-community/datetimepicker": "^8.3.0",
+    "@react-native-community/datetimepicker": "^8.2.0",
     "@react-navigation/bottom-tabs": "^7.2.0",
     "@react-navigation/native": "^7.0.14",
-    "expo": "~52.0.35",
+    "expo": "~52.0.36",
     "expo-blur": "~14.0.3",
-    "expo-constants": "~17.0.6",
+    "expo-constants": "~17.0.7",
     "expo-font": "~13.0.3",
     "expo-haptics": "~14.0.1",
     "expo-linking": "~7.0.5",
@@ -56,11 +56,10 @@
     "@types/react-native": "^0.72.8",
     "@types/react-test-renderer": "^18.3.1",
     "eslint": "^9.20.1",
-    "eslint-config-expo": "canary",
+    "eslint-config-expo": "^8.0.1",
     "eslint-plugin-react": "^7.37.4",
-    "expo-dev-client": "^5.0.12",
     "jest": "^29.7.0",
-    "jest-expo": "~52.0.4",
+    "jest-expo": "~52.0.5",
     "react-test-renderer": "^18.3.1",
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.24.1"
